New Feature Routes in task-manager-ui
